Added ForeignKeys and Index declarations. Added RoomCallback, where the default category is inserted.

+++ app/src/main/java/monakhova/bookmark/manager/presentation/mvi/model/Bookmark.kt
@@ -6,7 +6,7 @@ import monakhova.bookmark.manager.domain.models.IBookmark
  * Created by monakhova on 18.05.2020.
  */
 data class Bookmark(
-    override val id: Int = 1,
+    override val id: Long = 1,
     override val header: String,
     override val description: String,
     override val url: String,
